Moved to image segmentation package. WIP. still.
